pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = ""
authors = ["combitoo <combitoo@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
tortoise-orm = {extras = ["asyncmy"], version = "^0.20.0"}
jinja2 = "^3.1.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
fastapi-login = "^1.9.1"
aerich = "^0.7.2"



[tool.aerich]
tortoise_orm = "main.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
